Add `handleSubmit` method to `AddApplicationModel` and update `AddApplicationViewModel` to use it
* **AddApplicationModel.ts**
  - Add `handleSubmit` method to handle form submission
  - Move form submission logic from `AddApplicationViewModel` to `AddApplicationModel`
  - Import `Application` and `RootStore`
  - Update constructor to accept `rootStore` parameter

* **AddApplicationViewModel.ts**
  - Import `AddApplicationModel`
  - Replace `formData` with an instance of `AddApplicationModel`
  - Update methods to interact with `AddApplicationModel` instance
  - Remove form submission logic from `handleSubmit` method
